backend/requirements.txt
fastapi==0.116.1
openai==1.97.1
pypdf==6.0.0
python-dotenv==1.1.1
python-docx==1.2.0
requests==2.32.4
supabase==2.17.0
uvicorn==0.35.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pycryptodome==3.21.0
